Move core.utils in commons.utils
